requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
antlr4-python3-runtime==4.8
async-timeout==5.0.1
attrs==24.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
coloredlogs==15.0.1
configparser==5.1.0
datasets==3.1.0
ddddocr==1.4.7
dill==0.3.8
docker-pycreds==0.4.0
filelock==3.16.1
Flask==2.0.3
flatbuffers==23.5.26
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.9
GitPython==3.1.24
huggingface-hub==0.26.2
humanfriendly==10.0
hydra-core==1.1.1
idna==3.3
importlib_resources==6.4.5
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
numpy==1.23.5
olefile==0.47
omegaconf==2.1.1
onnxruntime==1.16.0
opencv-python-headless==4.8.0.76
packaging==23.1
pandas==2.0.3
pathtools==0.1.2
Pillow==8.4.0
promise==2.3
propcache==0.2.0
protobuf==3.19.1
psutil==5.8.0
pyarrow==17.0.0
python-dateutil==2.8.2
pytz==2024.2
PyYAML==6.0.1
requests==2.32.2
sentry-sdk==1.4.3
shortuuid==1.0.8
six==1.16.0
smmap==5.0.0
subprocess32==3.5.4
sympy==1.12
termcolor==1.1.0
torch==1.13.0+cu117
torchaudio==0.13.0
torchvision==0.14.0+cu117
tqdm==4.66.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.7
utils==1.0.2
wandb==0.12.6
Werkzeug==2.3.7
xxhash==3.5.0
yarl==1.15.2
yaspin==2.1.0
zipp==3.20.2
